requirements.txt
coverage
mock
nose
pea
sphinx
-e git://github.com/lacostej/nosyd.git#egg=nosyd
git+https://github.com/python-zk/kazoo.git
zc-zookeeper-static
